Remove AccountLocker description and update balance data types

+++ banking_core/database/setup.py
@@ -7,7 +7,7 @@ def setup_database(connection):
                             id SERIAL PRIMARY KEY,
                             number TEXT UNIQUE,
                             pin TEXT,
-                            balance INTEGER DEFAULT 0,
+                            balance NUMERIC(10, 2) DEFAULT 0,
                             failed_attempts INTEGER DEFAULT 0,
                             locked BOOLEAN DEFAULT FALSE
                         );''')
@@ -15,14 +15,14 @@ def setup_database(connection):
                             id SERIAL PRIMARY KEY,
                             account_number TEXT,
                             transaction_type TEXT,
-                            amount INTEGER,
+                            amount NUMERIC(10, 2),
                             date TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
                             FOREIGN KEY (account_number) REFERENCES card(number)
                         );''')
         cursor.execute('''CREATE TABLE IF NOT EXISTS daily_limits (
                             id SERIAL PRIMARY KEY,
                             account_number TEXT,
-                            daily_limit INTEGER,
+                            daily_limit NUMERIC(10, 2),
                             set_date TEXT,
                             changes_today INTEGER DEFAULT 0,
                             FOREIGN KEY (account_number) REFERENCES card(number)
